functions/inference/requirements.txt
# Flask for HTTP routing
Flask==3.0.0

# Firebase & Google Cloud
firebase-admin==6.2.0
google-cloud-firestore==2.13.0
google-cloud-storage==2.13.0
google-cloud-speech==2.21.0
google-cloud-translate==3.12.1
google-cloud-tasks==2.14.0

# Audio/Video Processing
ffmpeg-python==0.2.0
pydub==0.25.1

# Speaker Diarization
resemblyzer==0.1.1.dev0
scikit-learn==1.3.0
hdbscan==0.8.33

# Core ML dependencies
transformers==4.51.3
accelerate>=0.25.0
safetensors>=0.4.0
diffusers

# HuggingFace
huggingface-hub>=0.17.0
tokenizers>=0.14.0

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1
noisereduce>=3.0.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0

# Optional but recommended for performance
einops>=0.7.0